Changing constraint with not null, and a double constraint to allow for null in habitat.
Some of the constraint have not been harsh enough, even with a foreigh
key some data (few but some) had no value for stage or qal_id, so when
setting a unique constraint it did not work (we had duplicates, I found
this while testing insertions of multiple row of Vattican, they never
fished so much!). Inserting a constraint with a where clause is the
trick to avoid the problem with null values in constraint.