78: Develop integration script in shiny for recruitment, yellow and silver data
Added a column with reference to datasource for table, t_series_ser
(e.g. series), t_dataseries_das (glass yellow and silver recruitment
series), and t_biometry_bio. Note that all source have been set to 2019
for t_series_ser, and that source have been set to wgeel_2016 when last
date update was null, and to the datacall corresponding to the year of
change in other cases.

Task-Url: http://github.com/ices-eg/wg_WGEEL/issues/78