78: Develop integration script in shiny for recruitment, yellow and silver data
step 1, 2.1 OK
The new_data, updated_data, biometry, series are compared to database
and four sheets are created and displayed. data to insert, modified data
(whether they come from updated data or new_data so we don't  trust the
data provider blindly and pool entries from both sheets after
comparison). We also provide two sheets with highlight for change, these
only display the column that have changed (in any of the line). The
anti_join does not work fully, as discoved by Laurent and additional
step for equality is needed. This is done column per column, between the
database and excel entries. And the results are displayed in a file
showing only changes (the files are quite large so this should help the
end user to focus on the change in the database.  For updated data we
check that the id corresponding to year and das_ser_id matches the id
from the database, if it's not the case a warning is displayed in red.
Some changes brought to the datatable to decrease their size.

Task-Url: http://github.com/ices-eg/wg_WGEEL/issues/78